requirements.txt
apparmor==2.13.3
appdirs==1.4.3
application-utility==1.3.1
arandr==0.1.10
asn1crypto==0.24.0
autopep8==1.4.4
bauh==0.8.3
beautifulsoup4==4.8.0
bs4==0.0.1
btrfsutil==1.1.1
CacheControl==0.12.6
caffeine-ng==3.4.2
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
docopt==0.6.2
ewmh==0.1.6
future==0.18.2
gufw==19.10.0
html5lib==1.0.1
httplib2==0.9.2
hurry.filesize==0.9
idna==2.8
importlib-metadata==1.4.0
jeepney==0.4.2
keyring==10.6.0
keyrings.alt==3.0
layoutswitcherlib==0.8.2
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
LibAppArmor==2.13.3
louis==3.12.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==8.2.0
msgpack==0.6.2
netifaces==0.10.4
npyscreen==4.10.5
oauth==1.0.1
olefile==0.45.1
ordered-set==3.1.1
packaging==20.1
pacman-mirrors==4.16.2
pep517==0.8.1
pexpect==4.2.1
Pillow==6.2.1
power==1.4
progress==1.5
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.6.0
pwquality==1.4.2
pyaml==19.12.0
pycairo==1.19.1
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.74
pycurl==7.43.0.5
PyGObject==3.34.0
pyparsing==2.2.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyRFC3339==1.0
pysmbc==1.0.18
PySocks==1.7.0
python-dateutil==2.6.1
python-telegram-bot==12.4.2
python-xlib==0.26
pytoml==0.1.21
pytz==2019.3
pyxdg==0.26
PyYAML==5.3
reportlab==3.5.34
requests==2.22.0
requests-futures==1.0.0
requests-unixsocket==0.1.5
retrying==1.3.3
SecretStorage==3.1.2
setproctitle==1.1.10
simplejson==3.13.2
sip==4.19.21
six==1.11.0
soupsieve==1.9.2
stem==1.7.1
team==1.0
toml==0.10.0
tornado==6.0.3
torrequest==0.1.0
ufw==0.36
urllib3==1.25.3
wadllib==1.3.2
webencodings==0.5.1
yapf==0.28.0
zipp==3.0.0
